Change scale of PG objective to match other objectives.
Previously, when computing surrogate objective function for policy
gradient, we were summing over time instead of averaging. But, this
makes the scale of the function dependent on the number of time steps,
and makes it harder to choose weight values for the other objective
functions (entropy, policy evaluation, policy evaluation
regularization), since those in the code those ARE averaged over time.